refactor: Moved methods dealing with "nearest" files to `Util`
