+++ src/main/java/dev/jbang/Settings.java
@@ -15,6 +15,7 @@ public class Settings {
 	public static final String TRUSTED_SOURCES_JSON = "trusted-sources.json";
 	public static final String DEPENDENCY_CACHE_JSON = "dependency_cache.json";
 	public static final String CURRENT_JDK = "currentjdk";
+	public static final String JBANG_DOT_DIR = ".jbang";
 	public static final String BIN_DIR = "bin";
 	public static final String EDITOR_DIR = "editor";
