+++ src/main/java/dev/jbang/catalog/Catalog.java
@@ -263,12 +263,12 @@ public class Catalog {
 	}
 
 	private static Catalog findNearestCatalog(Path dir) {
-		Path catalogFile = CatalogUtil.findNearestFileWith(dir, JBANG_CATALOG_JSON, p -> true);
+		Path catalogFile = Util.findNearestFileWith(dir, JBANG_CATALOG_JSON, p -> true);
 		return catalogFile != null ? get(catalogFile) : null;
 	}
 
 	static Catalog findNearestCatalogWith(Path dir, Function<Catalog, Boolean> accept) {
-		Path catalogFile = CatalogUtil.findNearestFileWith(dir, JBANG_CATALOG_JSON, cat -> accept.apply(get(cat)));
+		Path catalogFile = Util.findNearestFileWith(dir, JBANG_CATALOG_JSON, cat -> accept.apply(get(cat)));
 		if (catalogFile == null) {
 			Path file = Settings.getUserImplicitCatalogFile();
 			if (Files.isRegularFile(file) && Files.isReadable(file) && accept.apply(get(file))) {
