+++ src/main/java/dev/jbang/cli/Catalog.java
@@ -37,7 +37,7 @@ abstract class BaseCatalogCommand extends BaseCommand {
 		} else {
 			if (catalogFile != null && Files.isDirectory(catalogFile)) {
 				Path defaultCatalog = catalogFile.resolve(dev.jbang.catalog.Catalog.JBANG_CATALOG_JSON);
-				Path hiddenCatalog = catalogFile.resolve(CatalogUtil.JBANG_DOT_DIR)
+				Path hiddenCatalog = catalogFile.resolve(Settings.JBANG_DOT_DIR)
 												.resolve(dev.jbang.catalog.Catalog.JBANG_CATALOG_JSON);
 				if (!Files.exists(defaultCatalog) && Files.exists(hiddenCatalog)) {
 					cat = hiddenCatalog;
