+++ src/main/java/dev/jbang/util/Util.java
@@ -40,6 +40,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Scanner;
+import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -1232,4 +1233,34 @@ public class Util {
 		}
 		return Paths.get("");
 	}
+
+	public static Path findNearestFileWith(Path dir, String fileName, Function<Path, Boolean> accept) {
+		Path result = findNearestLocalFileWith(dir, fileName, accept);
+		if (result == null) {
+			Path file = Settings.getConfigDir().resolve(fileName);
+			if (Files.isRegularFile(file) && Files.isReadable(file) && accept.apply(file)) {
+				result = file;
+			}
+		}
+		return result;
+	}
+
+	private static Path findNearestLocalFileWith(Path dir, String fileName, Function<Path, Boolean> accept) {
+		if (dir == null) {
+			dir = getCwd();
+		}
+		while (dir != null) {
+			Path file = dir.resolve(fileName);
+			if (Files.isRegularFile(file) && Files.isReadable(file) && accept.apply(file)) {
+				return file;
+			}
+			file = dir.resolve(Settings.JBANG_DOT_DIR).resolve(fileName);
+			if (Files.isRegularFile(file) && Files.isReadable(file) && accept.apply(file)) {
+				return file;
+			}
+			dir = dir.getParent();
+		}
+		return null;
+	}
+
 }
