+++ src/test/java/dev/jbang/cli/TestAlias.java
@@ -22,9 +22,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import dev.jbang.BaseTest;
+import dev.jbang.Settings;
 import dev.jbang.catalog.Alias;
 import dev.jbang.catalog.Catalog;
-import dev.jbang.catalog.CatalogUtil;
 import dev.jbang.util.Util;
 
 import picocli.CommandLine;
@@ -236,9 +236,9 @@ public class TestAlias extends BaseTest {
 		Path sub = Files.createDirectory(cwd.resolve("sub"));
 		Path testFile = sub.resolve("test.java");
 		Files.write(testFile, "// Test file".getBytes());
-		Path hiddenJBangPath = Paths.get(cwd.toString(), CatalogUtil.JBANG_DOT_DIR);
+		Path hiddenJBangPath = Paths.get(cwd.toString(), Settings.JBANG_DOT_DIR);
 		Files.createDirectory(hiddenJBangPath);
-		Files.createFile(Paths.get(cwd.toString(), CatalogUtil.JBANG_DOT_DIR, Catalog.JBANG_CATALOG_JSON));
+		Files.createFile(Paths.get(cwd.toString(), Settings.JBANG_DOT_DIR, Catalog.JBANG_CATALOG_JSON));
 		JBang jbang = new JBang();
 		new CommandLine(jbang).execute("alias", "add", "-f", cwd.toString(), "--name=name", testFile.toString());
 		assertThat(Files.isRegularFile(Paths.get(cwd.toString(), Catalog.JBANG_CATALOG_JSON)), is(false));
